Fix I18NSync-related issues 2836, 2894, and 2911.  Also includes generated output files, which are used by I18NSuite.  The selection of classloader to find properties files was changed to be more friendly to maven or other tools which may want to manipulate the classpath.
Note that this change makes the order of unique keys defined for property
files with duplicated keys (after non-identifier characters are replaced
with the underscores) deterministic.  If an application relied on the
previous behavior (which would generally be the same order if no new keys
were added but might change if any unrelated key were added or changed), the
de-duplicated keys may be different.  Hopefully the use of such keys in
properties files is rare :).

Issues: 2836, 2894, 2911
Patch by: jat
Review by: ajr, tobyr



git-svn-id: https://google-web-toolkit.googlecode.com/svn/releases/1.5@3756 8db76d5a-ed1c-0410-87a9-c151d255dfc7