Update patch set 2
Patch Set 2:

Well, doesn't that forces everyone using Git to make sure they rebase their changes on top of a git-svn-sync'd commit before pushing for review?
Force-push to a public repo is a admitted to be a Bad Practice, so maybe until the sync is reversed we'd better not merge in Git ("submit" changes in Gerrit) and simply abandon the review instead? (saying in the reason message it's been merged upstream and will sync later)

Patch-set: 2