Remove dependency from JSON-P on RPC
More specifically replace use of AsyncCallback with c.g.g.core.client.Callback,
and make AsyncCallback extend Callback so that existing code continues to
compile.
There's a small breaking change though in JsonpRequest#getCallback.

Creates the ApiChecker configuration file for 2.5 vs. 2.6 but doesn't
update the reference JAR, or build.xml to reference that new file. This
means that "ant apicheck" fails pending those changes (a rework of apicheck
is needed to be able to use a released SDK as reference instead of checking
in JARs to SVN).

Finally, updates a few uses of the raw-type AsyncCallback that no longer
compile due to type erasure (Java wants implementing classes to have a
onFailure(Object) method in addition to onFailure(Throwable)).

Change-Id: I73f88c42393d4c8536cab227bbebc11b6809f725