Update patch set 3
Patch Set 3:

FYI, in the process of reworking module dependencies, I've created a c.g.g.u.Timer module that depends only on c.g.g.u.Window, but unfortunately Window depends on DOM (for some Document methods). Note that Window also depends on HTTP (cyclic-dependency) for its URL and UrlBuilder classes.

So, without this patch (with HTTP depending on Timer), we'd be forced to bundle both DOM and HTTP in the same JAR. With this patch, we can put HTTP in the "core" JAR (alongside c.g.g.core.Core and XMLHttpRequest) or an "http" JAR, and Window and DOM together in a "DOM" JAR.

Of course, an alternative would be to break the dependency from Timer on Window in a similar way that this patch uses JSNI to no longer depend on Timer. We could then abandon this CL and bundle Timer and HTTP in "core", and Window and DOM in "dom", and have "rpc" only depend on "core", not "dom".

I haven't yet looked in details at c.g.g.junit.JUnit dependencies though, so maybe it needs Window and DOM too and this "core" vs. "dom" separation is pointless (with Elemental in the play though, I think it's not: you don't want c.g.g.dom.DOM to clutter your classpath and source-path when you use Elemental's own DOM classes; note that XMLHttpRequest is needed by GWT#runAsync, even though Elemental also provides its own version).

Patch-set: 3
Label: Code-Review=0