Disentangle module dependencies, breaks (most) cyclic dependencies.
There are basically 3 types of modules:
 - those that provide source or super-source
 - those that only provide configuration (deferred-binding, linker, etc.)
 - those that aggregate the above

The rule I followed is that non-source modules should try hard to not
import sources that are not already imported by their dependencies. They
use a <source path="null"/> to enforce that in some cases.
Additionally, configuration-only modules that are "private" (not intended
to being inherited directly by users) do not inherit any module, they just
assume the aggregate module that inherits them also inherits the required
dependencies and provides the sources.

To split the User module, RemoteService is made standalone, and a Cookies,
Random, Timer, and UI modules are introduced.
The c.g.g.user.HTTPRequest module is removed, as the HTTPRequest class that
it was tied to has been removed a long time ago.
Finally, a ValidationBase module has been added so that javax.validation
can be inherited for its API, without inheriting the whole c.g.g.validation
module.

Change-Id: I18ab07b3c31902c0af7fc4b0e716906265355329