Update patch set 1
Patch Set 1:

BTW, SignatureDumper (for javac.typemodel) outputs "method <init> Foo()V" for a constructor vs. "method Foo()V" for a method, whereas getJsniSignature uses "new()". It seems like either approach would work here, as long as the signature for a constructor is different from that for a method.

Patch-set: 1