Update patch set 1
Patch Set 1: Code-Review+1

Just a reminder: that's not enough as this is still in the Git history. The Git history needs to be rewritten (with git filter-branch) and force-pushed.

Before you do that, we also need to remove the reference JARs for the ApiChecker (we could download them from Central before running the tool for instance).

Patch-set: 1
Reviewer: Gerrit User 5030 <5030@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Code-Review=+1