Create change
Patch-set: 1
Change-id: I074f66ba12128b83ee2a98f3af84bafdc46746e2
Subject: Fixing a null reference exception in Mozilla. The problem occurs when running getBodyOffsetLeft/Top in an iframe that is not currently visible, such as via display:none. In this case, $wnd.getComputedStyle returns null. Fix this by null-checking the return value from getComputedStyle, and returning 0 for the body offset in this case. This should be safe to do since returning 0 is the default behavior defined in DOMImpl, and so is presumably used in other browsers.
Branch: refs/heads/master
Commit: 5d7d8121608fd9d74c90ed9d24e64c7829f61713