Update patch set 1
Patch Set 1:

I'm not sure what the best way to test this would be. Since the bug is browser specific, I don't know if GWTTestCase is appropriate. It also only appears when the application is running in an iframe, which I'm not sure how I'd set up in a GWTTestCase. Please let me know if you have any advice.

Patch-set: 1
Label: Code-Review=0