Update patch set 2
Patch Set 2:

I tried writing the following test:

  public void testGetBodyOffsets_insideHiddenIframe() {
    Document doc = Document.get();

    IFrameElement iframe = doc.createIFrameElement();
    iframe.getStyle().setDisplay(Display.NONE);
    doc.getBody().appendChild(iframe);

    assertEquals(0, iframe.getContentDocument().getBodyOffsetTop());
    assertEquals(0, iframe.getContentDocument().getBodyOffsetLeft());
  }

This almost got me what I want, but unfortunately is passes even without my change. I believe the reason is that getBodyOffsetTop calls getComputedStyle on $wnd, which is bound to the top-level window rather than the iframe window. I wrote some javascript and confirmed that getComputedStyle is expected to return non-null if I use the top-level window instead of the iframe window.

I think in order to write the test I would need to actually be executing the GWT code from within an iframe. Any ideas on how I can do/emulate this from a GWTTestCase? I could probably run JSNI in the test to rebind $wnd, but that sounds sketchy.

Patch-set: 2