Update patch set 2
Patch Set 2:

It's probably still not fully WAI. issue 4056 describes this as an issue in other browsers too, where Firefox fails with an exception and the other browsers have broken layouts. I think this will make Firefox behave the same as the other browsers, which will definitely be an improvement in some cases. I'm not sure how feasible it will be to completely fix the issue since the browser just isn't giving us any information in this case.

Patch-set: 2