Update patch set 2
Patch Set 2:

Sorry for not being clear before. 
I actually was referring to your testing attempt. You said
".. believe the reason is that getBodyOffsetTop calls getComputedStyle on $wnd, which is bound to the top-level window rather than the iframe window..".
I was questioning if doc.createIFrameElement().getContentDocument().getBodyOffsetTop() should really return the offset of the parent. That may be bug, perhaps it should return its own offset. 
(And if it was working correctly then you would have been successful in testing the behavior.)

Patch-set: 2