Update patch set 2
Patch Set 2:

It's not really clear to me what the expected result in this situation is, or even what the practical difference is between using the two different window objects. I'd be hesitant to change it without further exploring the effect. I think that's a little outside the scope of this change - would you prefer I submit this as is or wait for your question to be resolved?

Patch-set: 2
Reviewer: Gerrit User 5155 <5155@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Code-Review=0