Update patch set 1
Patch Set 1:

(4 comments)

W.r.t. to speedtracer: I was exploring using the speedtracer output but the format is too verbose; generating files that are extremely large (and we don't want that). The current version of speedtracer can not event load the output from one compile. (BTW the current version of speedtracer is broken and can not even load its own saved data, I submitted a fix for that)

One option is to use the current speedtracer instrumentation but adding a more reasonable alternative output format and adding a post processing script that allows filtering/aggregation to focus on the events of interest.

Patch-set: 1
Label: Code-Review=0