Update patch set 1
Patch Set 1:

(1 comment)

I am taking the shortcut with SurpressWarnings for the tests that are not written by GWT team but copied from Apache. I'm not planning to return back and fix those. They don't look like something we are actively maintaining and we might import and over-write with the newer versions later.

Patch-set: 1