Update patch set 2
Patch Set 2:

For refiewing this huge change, I suggest we first look at the APIs in c.g.g.i18n.shared and make sure we are happy with them, then look at some of the implementations of them, then the c.g.g.i18n.client compatibility code, and finally the generation code.  Sound good?

We still need to figure out best practices for using this in a multithreaded server, since the traditional approach of using static to hold a created instance isn't going to work (unless we make that instance just a delegate to a per-thread instance, which would be expensive without more information from the caller as every GWT.create would have to do that).  I have some ideas, but I think we can safely postpone that to a later change.

Patch-set: 2