Update patch set 2
Patch Set 2:

Ok.  For reviewing the APIs, Mark Davis or Shanjian Li might be good choices.

Can someone familiar with GWT make a CL from this change and run it through TAP to make sure it didn't break anything?  The only changes should be in any code that runs in the default locale, since some of those changed, but nobody should be doing that.

Patch-set: 2