Update patch set 2
Patch Set 2:

(7 comments)

Thanks for the review.

The ICU downloadable jars don't include the CLDR utilities.jar.  As for maven, I will leave that to someone else as maven and I don't get along :) -- if you can get the current jars via maven then that's fine with me.

In the past, we had to have local mods to ICU to get it to work with CLDR tools, but that hasn't been necessary for a while.

Patch-set: 2
Label: Code-Review=0