Update patch set 1
Patch Set 1:

On Mon, Jan 7, 2013 at 3:06 PM, Goktug Gokdogan <noreply-gerritcodereview-_6yca3-5C7TGMdxkHixTfA@google.com> wrote:

> catchExceptions is just for debugging not for this purpose and will cause no exceptions to be caught, including regular exceptions (even assertions); it will not report failures so basically test case will be not really working.
So there is no opt-out.

Good point. I guess we'd have to add one.
 
> Even in the reported issue, the guy debugged hours to see why it wasn't not called; which basically means it was important in their code to be called so he spent hours looking why it didn't happen.

My interpretation is that he didn't realize that the code under test was installing an uncaught exception handler. It was unexpected interference that caused GWTTestCase to stop working. The reason it took hours is because GWTTestCase didn't detect this interference and report it, so he's basically asking for better errors.

> At the end, it is a big surprise for the developer when it happens, no matter if we document it or not.

Yes, sometimes code being tested does surprising things. Part of the point of testing is to discover the surprises and report them so developers can decide what to do. If it's actually correct behavior then it can be declared.

So it seems like a test that uses GWT's uncaught exception handler should declare it explicitly somehow, since this is unusual behavior. (Much like a "regular" test shouldn't rely on test method order, but if it arranges this with the test infrastructure then then it's okay.)

Patch-set: 1
Label: Code-Review=0