Update patch set 1
Patch Set 1:

(Repost since Gerrit doesn't allow reply by email.)

On Mon, Jan 7, 2013 at 1:55 PM, Thomas Broyer wrote:

> It would mean generating a map (or an array indexed by
ordinal()) on the client-side, and I'm not sure how we'd do
it on the server-side. Maybe we should store those in the
serialization policy then?

For client-side we'd need another array included in the generated code for the enum serialization/deserialization, which could be omitted if the constants happened to be alphabetical so the ord() value could be used directly. For server-side we could probably get the name of each enum constant from the Java class, so we could calculate the array on first use.

But on second thought, that's better done as a separate change (if at all); the protocol was already very brittle and we're just making that visible.
 
I don't think we can easily test anything else (particularly not using the same enum definition on the client- vs. server-side).

Maybe construct two instances of JEnumType with the same name but different constants, and make sure they have different checksums? That should fail without the change and pass with it.

Patch-set: 1
Label: Verified=0