Update patch set 1
Patch Set 1:

Ok before moving forward, to summarize our requirements are:
 - detect and fail when uncaught handler is removed
 - ability to opt-out from GWTTestCase uncaught handlers

One thing that is still not clear is for the cases where the developer doesn't have access to code that adds an uncaught handler. I'm assuming we are not covering that until somebody comes with a use case for that.

Patch-set: 1
Label: Code-Review=0