Update patch set 5
Patch Set 5:

The "tools" directory hasn't been converted to Gerrit yet. I'll follow the directions you gave. It looks like we're using a patched version of 50.1 outside GWT and I don't see any evidence of 50.1.1 but a slight mismatch seems okay.

After I get tools updated, I think we should upgrade GWT to use ICU 50.1.1 as a separate patch. It seems like it should be okay to do before the maintenance release?

Also, there's a separate change pending (it got rolled back, but the plan is to reapply it) to upgrade the CLDR numeric constants to CLDR 21. Is that going to conflict? What's the relationship between ICU and CLDR anyway?

Patch-set: 5