Update patch set 4
Patch Set 4: Code-Review-1

I have seen this code before and questioned why it didn't have the fields in the first place. I think it is carefully crafted this way in order to be able to fully compile out logging code. Even the earlier version was not just keeping them in simple fields.

So I would not change it without a deeper analysis / discussion.

Patch-set: 4