Update patch set 4
Patch Set 4:

I tend to agree with Brian (and like Goktug I wondered about side-effects) but still find it annoying that we have a regression. And I'm not sure Roberto's analysis is right, in which case (if there's no negative side-effect) I'd favor applying this change.

If you don't define any custom Level, then getName() and intValue() will always be overridden and, according to issue 4893 (which I quickly verified seems to indeed be fixed), will thus be turned into abstract methods in the Level class. Following this, the fields won't be read any-longer and could thus be pruned.

This is where I'm not sure how the compiler works: will it prune fields that are only written but never read?

What if the default constructor (which is not part of the API) left the fields at their default value instead of setting them to "DUMMY" and -1 respectively?

If the above analysis is right, then people would have to pay the price for the almost-always-empty-fields only if they define (and use) custom levels.

Patch-set: 4
Reviewer: Gerrit User 5030 <5030@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Code-Review=0