Update patch set 6
Patch Set 6:

I removed the JDK7 calls -- not sure how that got by, since Eclipse was configured to enforce 1.6 compliance.

I'll make a separate patch just updating ICU4J, and the changes necessary to gwt and cldr-import to use it.  We can get that committed anytime, since it doesn't really change anything, and then this patch can rely on that.

Patch-set: 6