Update patch set 2
Patch Set 2: Code-Review-1

Wouldn't the proper way of doing this to add a configuration-property that the GWT-RPC will use in place of GWT.getModuleBaseURL() as the first argument to the generated RemoteServiceProxy-subclass constructor? It would thus also remove the need to setServiceEntryPoint on the instances: @RemoteServiceRelativePath would just work.

BTW, why does this change depend on I73ea4b8f ?

Patch-set: 2
Label: Code-Review=-1