Update patch set 4
Patch Set 4:

Goktug asked me to chime in and I honestly just don't remember if this was put in to make the logging code compile out - it seems likely since it is a strange pattern...

On the other hand, it's also likely that the compiler or some of the other code has since changed and it may no longer be necessary.  Easiest thing (that worked years ago at least) is to write some simple logging code (or use the logging example if it's still alive and kicking), disable logging, compile pretty and simply search for Level and Logger and make sure all mention of them is compiled out.

- Unnur

Patch-set: 4