Update patch set 2
Patch Set 2:

(3 comments)

Oops, I had some comments stuck as a "draft" in Gerrit. (Sorry about the delay.)

I asked outside the code review about the test failure. It turns out that RPCTypeCheckTest was failing all along, but we didn't notice because it isn't included in any test suite. I added it to RpcSuiteNoBrowser and confirmed that we need to catch AssertionFailedError instead of ArrayIndexOutOfBoundsException to make that test pass.

Patch-set: 2
Label: Code-Review=0