Update patch set 3
Patch Set 3:

The original problem isn't that we are interfering with the user handlers. It's that they are interfering with us, probably unintentionally. (That is, the user code was uninstalling our handler.) They are already running without any uncaught exception handler and we can't change that without making some tests fail that didn't fail before, because they were ignoring the errors.

One possibility would be to install the handler before gwtSetUp (just like before). After gwtSetUp, we can see if it's been replaced. If so, we replace it with our own and forward the call. But that still causes more failures than before.

I would like to actually clean up the mess, but I think we need to do more than just an assertion because the stack trace isn't good enough. (It doesn't contain the call to GWT.setUncaughtExceptionHandler() that caused the test to fail.) I'm working on a modification to your patch that I think will show the error more clearly.

Patch-set: 3
Reviewer: Gerrit User 5010 <5010@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Code-Review=0