Update patch set 1
Patch Set 1:

Okay, sounds fine then. But I don't understand the relationship between the ICU4J and CLDR libs in GWT. Should they be upgraded at the same time or is it okay to do it independently? Andrew Bachman has a change (not in Gerrit) to upgrade CLDR.

Patch-set: 1
Label: Verified=0