Update patch set 1
Patch Set 1:

The only thing ICU4J is used for in GWT is generating custom date/time formats from a skeleton, or for cldr-import.  It would be better to upgrade everything CLDR-related together, but I don't think it is a big deal.

Regarding upgrading CLDR, I strongly suggest waiting until the shared i18n APIs are done -- the cldr-import in trunk doesn't update everything, so you will wind up with a mix of CLDR versions.  Also, there typically have to be changes made to cldr-import to accommodate the new data files.

It could be done as part of the same CL if you don't mind the bigger churn, and that would save some effort both in updating GWT and from updating GWT apps (in past cases, there were a lot of internal tests that broke after a CLDR upgrade because they depended on particular localizations and you had to figure out if the test was checking some particular result they cared about or if they just cared that it was properly localized).  Either way, the first step is to get CLDR diffs and libraries updated in gwt-tools - that isn't used by anything but cldr-import, so you can do it without affecting anything else.

Patch-set: 1