Update patch set 5
Patch Set 5:

John: that's what I was suggesting yes.

Actually I'm puzzled, because GWT-RPC is not supposed to support being deployed that way.

I don't have a strong opinion on this since I'm not a GWT-RPC guy. However, I don't think we should push for including it in 2.5.1, I think we need to discuss it further: gather the use cases and find the best option or dismiss them (in favor of RequestFactory or anything else designed to support having the client and server running different "versions" of the "protocol data"; IIRC that "support" was added to deRPC only, not "standard" RPC, but maybe I'm wrong)

Patch-set: 5