Update patch set 5
Patch Set 5:

The main use case for GWT RPC being used from applications that start from the file system is phonegap.

If gwt users already have a big GWT application using GWT RPC, they tend to reuse most of the code, even the GWT RPC stuff. We always tell them to consider using AutoBeans or RequestFactory, but sometimes they don't want to.

GWT RPC can actually be deployed very well in this way, the only thing you have to do is keep your different server versions around. In gwt-phonegap we have been doing this for two years now and there are many applications out there which use that.

I would like to see this go into GWT 2.5.1 (if this is a valid change), since in gwt-phonegap we actually using JSNI to set the final field of the RemoteServiceProxy, which is a very bad hack to get this working.

Another way of dealing with this (which we also did in the past) is to override the doGetSerializationPolicy in RemoteServiceServlet, but this is very hacky as well.

Where should we discuss this?

Patch-set: 5
Reviewer: Gerrit User 5045 <5045@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Code-Review=0