Make sure a method and constructor cannot have the same signature.
Currently, a constructor has the same signature as a method with the same
name and arguments and a return type of 'void'. This causes issues in
UnifyAst which uses the signature as a key in a map. To make them different
the signature of a constructor no longer includes a return type and is annotated 
with the string "<init>"; the signature for the default constructor of the Foo
class will thus now be "Foo() <init>" instead of "Foo()V".

Fixes issue 7824

Change-Id: Iebcd7d0060bf7ac39c59e316b7d6dfb4c9727d7e