Stop using prefixed API's in AnimationScheduler by default. Firefox and Safari will use the Timer-based implementation. For Chrome we can use requestAnimationFrame because it's unprefixed starting in Chrome 24.
Since Chrome and Safari are in the same permutation, a future version of Safari
that adds the unprefixed API will automatically pick up requestAnimationFrame
once it's unprefixed. I attempted to avoid depending on any details of the
JavaScript API (it's mostly based on the Mozilla implementation) but can't
predict what the final standard will be, so it still seems risky. Perhaps we
should test specifically for Chrome?

Keep in mind that some apps built with GWT 2.5.1 will be around for a long time.

(Not tested yet.)

Change-Id: I3011dceab489871a5864eed1ece47ec850d82425