Update patch set 2
Patch Set 2:

The problem is that we have 2 ways of identifying an object: using it's proxy type in both case and either a client ID or server ID.

The client ID is scoped to a RequestContext and only used for a) EntityProxies created on the client-side (known as ephemeral ID), b) ValueProxies wherever they come from (ValueProxies don't have IDs, we're synthetizing ones for cross-reference in the request/response payload). For EntityProxies originating from the server, only the server ID is ever used (aka persistent ID).

When persisting an EntityProxy created on the client on the server, it's then given a server ID (transitions from ephemeral to persistent ID, and now has both client –so it can be matched back to the proxy on the client-side– and server –so it can be referenced in subsequent requests– IDs).

Now to our problem: the SimpleProxyId's hashCode is based on whether it has a client ID or not: if it has one, it's used as the hashCode, otherwise the serverId's hashCode is used. So, when we receive {"T":"type", "S":"serverId", "C":1} and {"T":"type", "S":"serverId"}, they're conceptually identical, but technically have different hashCodes. Because the SimpleProxyId is used as a key to the proxy in a map, those 2 IDs in the same response would map to different proxies (so one of them won't ever be populated; and because proxies are lazily created, but "finalized" only based on a specific list, one of those proxies won't be "finalized" either, and will generate a "don't cross the streams" exception when used in another RequestContext).

This patch makes sure we'd always use {"T":"type, "S":"serverId", "C":1} everywhere in the response payload.

BTW, SimpleProxyId#equals violates the contract, as it's not symmetric. This is another problem that we hopefully never run into.

Patch-set: 2
Reviewer: Gerrit User 5030 <5030@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Code-Review=0