Update patch set 2
Patch Set 2:

It looks like this part of the flow is not covered at all with unit tests. Is it possible to add some test cases so that we will know if it gets broken again?

Patch-set: 2
Reviewer: Gerrit User 5110 <5110@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Code-Review=0