Update patch set 4
Patch Set 4:

I think Gotug has a good point; since we're unlikely to see another bad bug in requestAnimationFrame, this is an exercise in closing the barn door after the horses are gone. So I'm okay with releasing GWT 2.5.1 without this patch.

About the general problem:

For someone releasing a GWT app that they won't be able to upgrade quickly (it may be software they're shipping so it's out of their hands), setting a single "conservative mode" flag seems like the best option.

But I was actually thinking about a different case, where someone is reacting to a new browser bug by setting a flag, recompiling, and redeploying. It would be a shame to have to disable a feature in all browsers because of a browser bug in one of them, or to degrade all features because of a bug in one of them. So that's why I was thinking of having feature and browser-specific flags.

But actually, for a sysadmin, having to recompile a GWT app and do a release to make a change isn't a great workflow either. It would make more sense to have flags that can be set from the server at runtime so that they can be hooked into the administrator's dashboard.

So maybe we should have runtime options that come from the HTML page or a separate server-side file?

Patch-set: 4