Update patch set 4
Patch Set 4:

No, just make it a regular deferred binding property.  It can be set to a default value so it doesn't change the number of permutations for most people (and we can make it conservative or not), and we already support having the server set properties that can be picked up by the property provider (ie, locale), and can even be used with server-side script selection.  No need to invent new mechanisms.

Patch-set: 4