Prevents GWTTestCase to become zombie when communication with backend fails.
Before this patch, when GWTTestCase fails to communicate with JUnitShell
for any reason (e.g. test case thrown a js exception which is not serializable),
then it was retrying indefinitely. This can case to hang until the test case timeouts. 

This patch changes the default for maximum retries and also provides an alternative
lightweight communication to JUnitShell to report something went wrong.

Change-Id: Ie02ea83e1363483c3bf2ca2dc5be6d9344cde2cf