Deprecates com.google.gwt.benchmark.*.
This didn't get enough adaption externally or internally. 
The way it's implemented requires special support from GWTTestCase
infrastructure. Also, we are not going have an enviroment soon
to automate benchmarks written on this any time soon.

If there is enough demand later, parts of this code can be
resurrected to build a new benchmark system.

Change-Id: I51fe1c9e73e39df745e71bf1fbf459fd6cc78f7f