Reduce devmode memory usage by not caching bytecode.
Instead of caching bytecode in a byte[] array in a TypeData in
a CompiledClass, read it again from the DiskCache. It looks like
the bytecode is only used once per compile, so the increased
memory usage isn't worth the speedup. (Also, the OS typically
caches files.)

Testing: using our largest GWT app, this reduced memory consumption
from 1170k (at idle after a gc) to 650k. The time for the first
reload went from 1:34 to 1:31 and the second reload went from 4 to
5 seconds (measured up to the log message that the module is loaded).
This seems well within the margin of error.

Change-Id: I2306c8873cd4651de09ae61bf3f13b64fc4cb7a0