Update patch set 1
Patch Set 1:

I don't intend any change in semantics; I'm just documenting the preexisting constraints and improving performance.

The only call to Annotations.setParent() was in JRealClass, and the "annotations" field doesn't escape, so it seems pretty clear that this inheritance mechanism is only used in JRealClass.

I haven't traced the dataflow all the way back, so it's possible that we're getting the effect of inheritance via copying somewhere, but it's looking unlikely.

Patch-set: 1