Update patch set 1
Patch Set 1: Code-Review-1

Can we try to move away from storing (large) binaries in the repo? The JARs are available right from Central, and the createApiCheckerReferenceJars.sh processing is only there to make the JARs smaller *because* we commit them (they probably speed up the api-checker initialization a bit too, but that should be negligible).
Ideally I'd like all those "reference files" to be stripped out of the Git history (same as plugins/) when we make the final move to "Git as the source of truth".
I'm effectively asking that we revert issue 3445, because we now deploy JARs to Central so it's easier to download them (we can even download them automatically when building, based on the version we want to compare against)

Patch-set: 1