Update patch set 1
Patch Set 1:

We need to think about how the alternative will work with our internal build system, we cannot just download code from central, so perhaps we can keep an internal reference but that is a different story.
In the meantime I cannot submit code because I'm removing some APIs and api-checker is failing. What are you suggesting as alternative that will work right now?

Patch-set: 1