Update patch set 1
Patch Set 1: -Code-Review

If it's blocking you, then I'd say go ahead, and we'll fix this later. I created issue 8031 to track it.

My knowledge of api-checker is a bit too weak for me to approve this review (for example, why did you add the TypedArraysFactory super-source implementation to the excludedFiles_old?), but I removed my -1.

Patch-set: 1