Update patch set 1
Patch Set 1:

Thanks. It looks we need to get rid of this history no matter if the new jars are there or not. So this patch should not do much harm.
My knowledge of api-checker is also very weak. I just tried to make sense from comments and previous patches.
I created excludedFiles_old section from the old config file's excludedFiles_new section. So that is why TypedArraysFactory is there; it was excluded for 2.5 in the old config.

Patch-set: 1