Update patch set 4
Patch Set 4:

I'm not sure this is an improvement. The implementation in each visitor seems pretty trivial and if each visitor has its own private fields, you can do a usage search and the results are easy to follow. If they share an implementation in a common base class then the usages will be intermingled and it's harder to follow.

Patch-set: 4